% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dona.R
\name{dona_pipeline_path}
\alias{dona_pipeline_path}
\title{Find the installed DONA pipeline}
\usage{
dona_pipeline_path()
}
\value{
A character string containing the pipeline folder.
}
\description{
Returns the folder containing the pipeline scripts included with DONA.
}
