% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dona.R
\name{dona_run}
\alias{dona_run}
\title{Run DONA inference}
\usage{
dona_run(project_folder = ".")
}
\arguments{
\item{project_folder}{Folder created by \code{\link[=dona_copy_pipeline]{dona_copy_pipeline()}}.}
}
\value{
TRUE invisibly after completion.
}
\description{
Runs the main R inference workflow using the config.R file inside the selected project folder.
CellPhoneDB is run separately because it uses Python.
}
